Fixed bug in multiline arguments
The check to see if the opening parenthesis is in the same line as
the current argument was checking if there was an opening parenthesis in
the current line at the column position as the real opening parenthesis
of the current function argument list. This would cause functions like
this to be parsed as if they were single line function arguments:
longfuncname(
    funcname(1),
    15,
    null
)
When the cursor is anywhere in the line of the first argument,
    getline('.')[rightup[2]-1]=='('
would be true, because both opening parenthesis are in the same column.
Checking for
if line('.')==rightup[1]
is more straightforward.